refactor(knockdog): dogschool 도메인 중심으로 fsd 구조 개선
기존의 분산된 features (search-filter, dogschool-list)를 단일 dogschool feature로 통합하여 도메인 응집도 향상
