--- apps/knockdog/src/entities/search-filter/index.ts
@@ -1,1 +1,0 @@
-export * from './config/constant';
