--- apps/knockdog/src/features/dogschool-list/index.ts
@@ -1,1 +1,0 @@
-export * from './ui/DogSchoolCard';
