+++ apps/knockdog/src/features/dogschool/index.ts
@@ -1,0 +1,1 @@
+export * from './ui/DogSchoolList';
