+++ apps/knockdog/src/features/dogschool/ui/CardClipDefs.tsx
