+++ apps/knockdog/src/features/dogschool/ui/DogSchoolCard.tsx
