+++ apps/knockdog/src/features/dogschool/ui/DogSchoolImageSlider.tsx
