--- apps/knockdog/src/features/search-filter/index.ts
@@ -1,1 +1,0 @@
-export { FilterBottomSheet } from './ui/FilterBottomSheet';
