--- apps/knockdog/src/widgets/dogschool-list/index.ts
@@ -1,2 +1,0 @@
-// Public API
-export { DogSchoolListWidget } from './ui/DogSchoolListWidget';
